More lingering typeID type changes
Also beefed up the unit tests to catch this error

+++ aba-industry/industry-fetcher/src/main/java/com/aba/industry/fetch/client/impl/StaticDataExportBlueprintYamlService.java
@@ -98,7 +98,7 @@ public class StaticDataExportBlueprintYamlService implements BuildRequirementsPr
                                         .getValue();
         }
 
-        final long manuBpTypeId = productionBp.getBlueprintTypeID();
+        final int manuBpTypeId = productionBp.getBlueprintTypeID();
 
         //<editor-fold desc="Possible base bp lambda">
         possibleMatch = blueprints.getBlueprints()
